Enabled the images for each cell
The table needs more structure to it, so I enabled the cell image for each cell. Then, I changed the color to AEAEAE. This stays more in theme with our game so far. I tried to add a background to it, but every time I tried to implement it, It disabled my ability to drag and drop.